+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+- Use pre-built worker via CDN by MarpWorker's default ([#2](https://github.com/marp-team/marp-react/pull/2))
+
 ## v0.0.1 - 2019-02-12
 
 - Initial release.